rtx: collect per-leaf emissive surfaces with pvs data
